Fix dask / distributed installation
The previous run failed to run the Dask tests

https://github.com/rapidsai/dask-upstream-testing/actions/runs/13164022831/job/36739676950#step:9:392

```
ImportError while loading conftest '/__w/dask-upstream-testing/dask-upstream-testing/dask/dask/conftest.py'.
_pytest.pathlib.ImportPathMismatchError: ('dask.conftest', '/pyenv/versions/3.10.16/lib/python3.10/site-packages/dask/conftest.py', PosixPath('/__w/dask-upstream-testing/dask-upstream-testing/dask/dask/conftest.py'))
```

It looks like the local dask was uninstalled when we intalled distributed. Most likely,
this was because the `--depth=1` we cloned the repository with broke dask's local
version system (using git tags). This PR clones with a larger depth.