Source env in the right script
Second attempt at fixing https://github.com/rapidsai/dask-upstream-testing/pull/14

That failed previously, because we sourced the uv env in the childe
`setup.sh` script, which can't modify the PATH of the paraent `./run.sh`
script.

So we do it in `install.sh`, which is the only place using `uv`.