Source uv in run.sh (#20)
Previous commits installed our packages into a virtualenv. That venv was
activated in `install.sh`, but wasn't in `test.sh`.

This moves the venv management to the parent `run.sh`. Then it should be
activate for both `install.sh` and `test.sh`.

The change to the `cuml` pytest call is from local testing reveling that
it might be necessary? I'm pretty sure I had a successful run without
the `--import-mode`, but now I was seeing errors about `E
AttributeError: module 'dask' has no attribute 'config'`, which
indicates that we had a directory named `dask` on our import path that
wasn't the actual dask package.

Fixes the failure at
https://github.com/rapidsai/dask-upstream-testing/actions/runs/13444080870/job/37565269167