Removed import-mode for cuml tests
Previous versions of this test runner needed `--import-mode=importlib`
to avoid the `cuml/tests/dask` directory from shadowing the `dask`
module. Now that we don't `cd` into the test directory, it seems
we're able to remove the `--import-mode` argument.

`--import-mode=importlib` was causing some other tests to fail with
serialization issues. Top-level functions defined in the test file would
be given a `__module__` of `cuml.tests`, which isn't a real module that's
importable on workers (which aren't usings pytest's special import
mechanism).

Closes https://github.com/rapidsai/dask-upstream-testing/issues/23