Add configurable Dask branch to cron workflow (#45)
This adds a new input to the cron workflow that allows a user to specify
a Dask branch or tag to test. The default is `main`, which is unchanged
from today.

We'll use this to test against a specific version of Dask version in
preparation for updating the pin in rapids-dask-dependency.