Temporarily hardcode released dask version (#46)
In https://github.com/rapidsai/dask-upstream-testing/pull/45 we tried
adding a workflow_dispatch parameter to set the Dask version. It didn't
make it all the way to the test.sh script; the shared-workflow setup
might not allow passing through variables.

So instead, we'll temporarily hardcode the dask version we test against.
After the release we'll revert this change.