Upper bound on RAPIDS libraries (#47)
We're preparing to update the version of dask used in
rapids-dask-dependency for the 2025.6.0 release.

This PR

- pins to dask==2025.4.1, the target version for rapids-dask-dependency
- Puts an upper bound of `<25.8.0a0` for rapids libraries, to ensure
that we (temporarily) get the 25.6.0 alphas

This ensures that we're testing what we'll release against.

We'll want to relax both of these pins once
rapids-dask-dependency is updated. Downstream projects' CIs will be
testing against the released version dask==2025.4.1 by that point. And
since the 2025.6.0 release will be using dask==2025.4.1, we don't need
to worry about testing it against `main`.