Pin to python 3.12 in matrix filter
This updates our matrix filter to select python 3.12.

Some dependencies (hdbscan ATM) don't have 3.13 wheels so we aren't able
to bump to that yet.

I think this is also going to result in the cuda 11 version not being
tested anymore (since it used python 3.10 IIRC). We were planning on dropping
that build anyway.