Try setting UV_LINK_MODE=copy
Our CI output currently contains messages like

    warning: Failed to hardlink files; falling back to full copy

e.g. https://github.com/rapidsai/dask-upstream-testing/actions/runs/15792973681/job/44521325602#step:11:435
This makes it hard to search for the string "failed". A previous attempt
at a fix set this in the `env` mapping but that apparently didn't work.
This tries setting it in `GITHUB_ENV` and hoping that it flows through
to the test script.